# Conflicts:
#	redbot/core/commands/commands.py
#	redbot/core/core_commands.py

+++ redbot/cogs/streams/streams.py
@@ -7,7 +7,6 @@ from redbot.core.utils.chat_formatting import escape, pagify
 
 from .streamtypes import (
     HitboxStream,
-    MixerStream,
     PicartoStream,
     Stream,
     TwitchStream,
@@ -40,7 +39,7 @@ log = logging.getLogger("red.core.cogs.Streams")
 class Streams(commands.Cog):
     """Various commands relating to streaming platforms.
 
-    You can check if a Twitch, YouTube, Picarto or Mixer stream is
+    You can check if a Twitch, YouTube or Picarto stream is
     currently live.
     """
 
@@ -79,6 +78,10 @@ class Streams(commands.Cog):
 
         self._ready_event: asyncio.Event = asyncio.Event()
         self._init_task: asyncio.Task = self.bot.loop.create_task(self.initialize())
+
+    async def red_delete_data_for_user(self, **kwargs):
+        """ Nothing to delete """
+        return
 
     def check_name_or_id(self, data: str) -> bool:
         matched = self.yt_cid_pattern.fullmatch(data)
@@ -141,11 +144,14 @@ class Streams(commands.Cog):
                     '2. Click "Manage" on your application.\n'
                     '3. Click on "New secret".\n'
                     "5. Copy your client ID and your client secret into:\n"
-                    "`[p]set api twitch client_id <your_client_id_here> "
-                    "client_secret <your_client_secret_here>`\n\n"
-                    "Note: These tokens are sensitive and should "
-                    "only be used in a private channel "
+                    "{command}"
+                    "\n\n"
+                    "Note: These tokens are sensitive and should only be used in a private channel "
                     "or in DM with the bot."
+                ).format(
+                    command="`[p]set api twitch client_id {} client_secret {}`".format(
+                        _("<your_client_id_here>"), _("<your_client_secret_here>")
+                    )
                 )
                 if notified_owner_missing_twitch_secret is False:
                     await send_to_owners_with_prefix_replaced(self.bot, message)
@@ -220,15 +226,9 @@ class Streams(commands.Cog):
         await self.check_online(ctx, stream)
 
     @commands.command()
-    async def hitbox(self, ctx: commands.Context, channel_name: str):
-        """Check if a Hitbox channel is live."""
+    async def smashcast(self, ctx: commands.Context, channel_name: str):
+        """Check if a smashcast channel is live."""
         stream = HitboxStream(name=channel_name)
-        await self.check_online(ctx, stream)
-
-    @commands.command()
-    async def mixer(self, ctx: commands.Context, channel_name: str):
-        """Check if a Mixer channel is live."""
-        stream = MixerStream(name=channel_name)
         await self.check_online(ctx, stream)
 
     @commands.command()
@@ -240,7 +240,7 @@ class Streams(commands.Cog):
     async def check_online(
         self,
         ctx: commands.Context,
-        stream: Union[PicartoStream, MixerStream, HitboxStream, YoutubeStream, TwitchStream],
+        stream: Union[PicartoStream, HitboxStream, YoutubeStream, TwitchStream],
     ):
         try:
             info = await stream.is_online()
@@ -250,17 +250,15 @@ class Streams(commands.Cog):
             await ctx.send(_("That channel doesn't seem to exist."))
         except InvalidTwitchCredentials:
             await ctx.send(
-                _(
-                    "The Twitch token is either invalid or has not been set. See "
-                    "`{prefix}streamset twitchtoken`."
-                ).format(prefix=ctx.clean_prefix)
+                _("The Twitch token is either invalid or has not been set. See {command}.").format(
+                    command=f"`{ctx.clean_prefix}streamset twitchtoken`"
+                )
             )
         except InvalidYoutubeCredentials:
             await ctx.send(
                 _(
-                    "The YouTube API key is either invalid or has not been set. See "
-                    "`{prefix}streamset youtubekey`."
-                ).format(prefix=ctx.clean_prefix)
+                    "The YouTube API key is either invalid or has not been set. See {command}."
+                ).format(command=f"`{ctx.clean_prefix}streamset youtubekey`")
             )
         except APIError:
             await ctx.send(
@@ -279,7 +277,7 @@ class Streams(commands.Cog):
 
     @commands.group()
     @commands.guild_only()
-    @checks.mod()
+    @checks.mod_or_permissions(manage_channels=True)
     async def streamalert(self, ctx: commands.Context):
         """Manage automated stream alerts."""
         pass
@@ -307,15 +305,10 @@ class Streams(commands.Cog):
         """Toggle alerts in this channel for a YouTube stream."""
         await self.stream_alert(ctx, YoutubeStream, channel_name_or_id)
 
-    @streamalert.command(name="hitbox")
-    async def hitbox_alert(self, ctx: commands.Context, channel_name: str):
-        """Toggle alerts in this channel for a Hitbox stream."""
+    @streamalert.command(name="smashcast")
+    async def smashcast_alert(self, ctx: commands.Context, channel_name: str):
+        """Toggle alerts in this channel for a Smashcast stream."""
         await self.stream_alert(ctx, HitboxStream, channel_name)
-
-    @streamalert.command(name="mixer")
-    async def mixer_alert(self, ctx: commands.Context, channel_name: str):
-        """Toggle alerts in this channel for a Mixer stream."""
-        await self.stream_alert(ctx, MixerStream, channel_name)
 
     @streamalert.command(name="picarto")
     async def picarto_alert(self, ctx: commands.Context, channel_name: str):
@@ -405,17 +398,16 @@ class Streams(commands.Cog):
             except InvalidTwitchCredentials:
                 await ctx.send(
                     _(
-                        "The Twitch token is either invalid or has not been set. See "
-                        "`{prefix}streamset twitchtoken`."
-                    ).format(prefix=ctx.clean_prefix)
+                        "The Twitch token is either invalid or has not been set. See {command}."
+                    ).format(command=f"`{ctx.clean_prefix}streamset twitchtoken`")
                 )
                 return
             except InvalidYoutubeCredentials:
                 await ctx.send(
                     _(
                         "The YouTube API key is either invalid or has not been set. See "
-                        "`{prefix}streamset youtubekey`."
-                    ).format(prefix=ctx.clean_prefix)
+                        "{command}."
+                    ).format(command=f"`{ctx.clean_prefix}streamset youtubekey`")
                 )
                 return
             except APIError:
@@ -431,9 +423,9 @@ class Streams(commands.Cog):
         await self.add_or_remove(ctx, stream)
 
     @commands.group()
-    @checks.mod()
+    @checks.mod_or_permissions(manage_channels=True)
     async def streamset(self, ctx: commands.Context):
-        """Set tokens for accessing streams."""
+        """Manage stream alert settings."""
         pass
 
     @streamset.command(name="timer")
@@ -452,7 +444,6 @@ class Streams(commands.Cog):
     @checks.is_owner()
     async def twitchtoken(self, ctx: commands.Context):
         """Explain how to set the twitch token."""
-
         message = _(
             "To set the twitch API tokens, follow these steps:\n"
             "1. Go to this page: https://dev.twitch.tv/dashboard/apps.\n"
@@ -461,11 +452,15 @@ class Streams(commands.Cog):
             "select an Application Category of your choosing.\n"
             "4. Click *Register*.\n"
             "5. Copy your client ID and your client secret into:\n"
-            "`{prefix}set api twitch client_id <your_client_id_here> "
-            "client_secret <your_client_secret_here>`\n\n"
+            "{command}"
+            "\n\n"
             "Note: These tokens are sensitive and should only be used in a private channel\n"
             "or in DM with the bot.\n"
-        ).format(prefix=ctx.clean_prefix)
+        ).format(
+            command="`{}set api twitch client_id {} client_secret {}`".format(
+                ctx.clean_prefix, _("<your_client_id_here>"), _("<your_client_secret_here>")
+            )
+        )
 
         await ctx.maybe_send_embed(message)
 
@@ -483,10 +478,14 @@ class Streams(commands.Cog):
             "3. Set up your API key \n"
             "(see https://support.google.com/googleapi/answer/6158862 for instructions)\n"
             "4. Copy your API key and run the command "
-            "`{prefix}set api youtube api_key <your_api_key_here>`\n\n"
+            "{command}\n\n"
             "Note: These tokens are sensitive and should only be used in a private channel\n"
             "or in DM with the bot.\n"
-        ).format(prefix=ctx.clean_prefix)
+        ).format(
+            command="`{}set api youtube api_key {}`".format(
+                ctx.clean_prefix, _("<your_api_key_here>")
+            )
+        )
 
         await ctx.maybe_send_embed(message)
 
@@ -695,6 +694,8 @@ class Streams(commands.Cog):
                         continue
                     for message in stream._messages_cache:
                         with contextlib.suppress(Exception):
+                            if await self.bot.cog_disabled_in_guild(self, message.guild):
+                                continue
                             autodelete = await self.config.guild(message.guild).autodelete()
                             if autodelete:
                                 await message.delete()
@@ -706,6 +707,8 @@ class Streams(commands.Cog):
                     for channel_id in stream.channels:
                         channel = self.bot.get_channel(channel_id)
                         if not channel:
+                            continue
+                        if await self.bot.cog_disabled_in_guild(self, channel.guild):
                             continue
                         ignore_reruns = await self.config.guild(channel.guild).ignore_reruns()
                         if ignore_reruns and is_rerun:
